+++ example/test.py
@@ -33,11 +33,12 @@ array_map = ArrayMap.read(filename=ARRAY_MAP_PATH)  # read in array map
 array_map.set_band(band=BAND)
 
 obs_log = ObsLog.read_folder(folder=OBS_LOG_FD)  # read in obs_fft log
+PARALLEL = True
 
 flat_result = reduce_skychop(
         flat_header=FLAT_HEADER, data_dir=DATA_PATH, write_dir=WRITE_PATH,
         array_map=array_map, obs_log=obs_log, pix_flag_list=[],
-        parallel=True, return_ts=False, return_pix_flag_list=True,
+        parallel=PARALLEL, return_ts=False, return_pix_flag_list=True,
         table_save=True, plot=True, plot_ts=True, reg_interest=REG_INTEREST,
         plot_flux=True, plot_show=False, plot_save=True)
 flat_flux, flat_err, flat_wt, pix_flag_list = flat_result
@@ -45,7 +46,7 @@ flat_flux, flat_err, flat_wt, pix_flag_list = flat_result
 zobs_result = reduce_zobs(
         data_header=DATA_HEADER, data_dir=DATA_PATH, write_dir=WRITE_PATH,
         array_map=array_map, obs_log=obs_log, pix_flag_list=pix_flag_list,
-        flat_flux=flat_flux, flat_err=flat_err, parallel=True, stack=True,
+        flat_flux=flat_flux, flat_err=flat_err, parallel=PARALLEL, stack=True,
         do_desnake=False, ref_pix=REF_PIX, do_smooth=False, do_ica=True,
         spat_excl=[0, 3], return_ts=False, return_pix_flag_list=False,
         table_save=True, plot=True, plot_ts=True,
@@ -63,7 +64,7 @@ zobs_result = reduce_zobs(
 # flat_result = reduce_skychop(
 #         flat_header=FLAT_HEADER,  data_dir=DATA_PATH, write_dir=WRITE_PATH,
 #         array_map=array_map, obs_log=obs_log, pix_flag_list=[],
-#         parallel=True, return_ts=False, return_pix_flag_list=True,
+#         parallel=PARALLEL, return_ts=False, return_pix_flag_list=True,
 #         table_save=False, plot=False, plot_ts=False, reg_interest=REG_INTEREST,
 #         plot_flux=True, plot_show=False, plot_save=True)
 # flat_flux, flat_err, flat_wt, pix_flag_list = flat_result
@@ -72,7 +73,7 @@ zobs_result = reduce_zobs(
 #         data_header=DATA_HEADER, data_dir=DATA_PATH, write_dir=WRITE_PATH,
 #         array_map=array_map, obs_log=obs_log, is_flat=False,
 #         pix_flag_list=pix_flag_list, flat_flux=flat_flux, flat_err=flat_err, cross=True,
-#         parallel=True, do_desnake=False, ref_pix=[1, 11],
+#         parallel=PARALLEL, do_desnake=False, ref_pix=[1, 11],
 #         do_smooth=False, do_ica=False, spat_excl=None, return_ts=False,
 #         return_pix_flag_list=True, table_save=True, plot=True,
 #         plot_ts=True, reg_interest=None, plot_flux=True, plot_show=False,
@@ -82,7 +83,7 @@ zobs_result = reduce_zobs(
 #         data_header=DATA_HEADER, data_dir=DATA_PATH, write_dir=WRITE_PATH,
 #         array_map=array_map, obs_log=obs_log, is_flat=False,
 #         pix_flag_list=pix_flag_list, flat_flux=flat_flux, flat_err=flat_err,
-#         parallel=True, do_desnake=False, ref_pix=[1, 11], do_smooth=True,
+#         parallel=PARALLEL, do_desnake=False, ref_pix=[1, 11], do_smooth=True,
 #         do_ica=False, spat_excl=(0, 5), return_ts=False,
 #         return_pix_flag_list=True, table_save=True, plot=True, plot_ts=False,
 #         reg_interest=None, plot_flux=True, plot_show=False, plot_save=True,
@@ -91,7 +92,7 @@ zobs_result = reduce_zobs(
 # eval_result = eval_performance(
 #         data_header=DATA_HEADER, data_dir=DATA_PATH, write_dir=WRITE_PATH,
 #         array_map=array_map, obs_log=obs_log, pix_flag_list=[],
-#         parallel=True, return_ts=False, table_save=True, plot=True,
+#         parallel=PARALLEL, return_ts=False, table_save=True, plot=True,
 #         plot_ts=True, reg_interest=REG_INTEREST, plot_psd=True, plot_specgram=True,
 #         plot_flux=True, plot_show=False, plot_save=True)